+++ YACReaderLibrary/properties_dialog.h
@@ -19,6 +19,10 @@ class QComboBox;
 // class YACReaderBusyWidget;
 class QToolButton;
 
+namespace YACReader {
+class CoverLabel;
+}
+
 #include "comic_db.h"
 
 class PropertiesDialog : public QDialog
@@ -33,7 +37,7 @@ private:
     QTabWidget *tabBar;
 
     QWidget *coverBox;
-    QLabel *cover;
+    YACReader::CoverLabel *cover;
     QScrollArea *sa;
 
     QWidget *generalInfoBox;
@@ -113,11 +117,12 @@ private:
     QPushButton *previousButton;
     QPushButton *restoreButton; //??
 
-    QPixmap coverImage;
-
     QToolButton *showPreviousCoverPageButton;
     QToolButton *showNextCoverPageButton;
     QLabel *coverPageNumberLabel;
+
+    QToolButton *resetCoverButton;
+    QToolButton *loadCustomCoverImageButton;
 
     void createTabBar();
     void createCoverBox();
@@ -143,6 +148,8 @@ private:
     float coverSizeRatio;
     bool updated;
     QString originalCoverSize;
+
+    QImage customCover;
 
 public:
     PropertiesDialog(QWidget *parent = nullptr);
@@ -170,7 +177,13 @@ public slots:
     void setSize(float size);
     void loadNextCover();
     void loadPreviousCover();
+    void resetCover();
+    void loadCustomCoverImage();
+    void loadCustomCoverImageFromPath(const QString &path);
+    void setCoverPage(int pageNumber);
     bool close();
+    void updateCoverBoxForMultipleComics();
+    void updateCoverBoxForSingleComic();
 
 signals:
     void coverChangedSignal(const ComicDB &comic);