+++ common/comic_db.h
@@ -84,8 +84,6 @@ public:
 
     QVariant comicVineID; // string
 
-    QImage cover;
-
     QVariant lastTimeOpened; // integer/date
     QVariant coverSizeRatio; // h/w
     QVariant originalCoverSize; // string "WxH"
@@ -188,8 +186,6 @@ public:
     Q_PROPERTY(QVariant notes MEMBER notes CONSTANT)
 
     Q_PROPERTY(QVariant comicVineID MEMBER comicVineID CONSTANT)
-
-    Q_PROPERTY(QImage cover MEMBER cover CONSTANT)
 
     Q_PROPERTY(QVariant lastTimeOpened MEMBER lastTimeOpened CONSTANT)