+++ common/yacreader_global_gui.cpp
@@ -2,6 +2,8 @@
 
 #include <QtCore>
 #include <QAction>
+#include <QImageReader>
+#include <QFileDialog>
 
 using namespace YACReader;
 
@@ -98,4 +100,48 @@ QAction *YACReader::actionWithCustomIcon(const QIcon &icon, QAction *action)
 QPixmap YACReader::hdpiPixmap(const QString &file, QSize size)
 {
     return QIcon(file).pixmap(size);
+}
+
+QString YACReader::imageFileLoader(QWidget *parent)
+{
+    QString supportedImageFormatsString;
+    for (const QByteArray &format : QImageReader::supportedImageFormats()) {
+        supportedImageFormatsString += QString("*.%1 ").arg(QString(format));
+    }
+
+    return QFileDialog::getOpenFileName(parent, QObject::tr("Select custom cover"), QDir::homePath(), QObject::tr("Images (%1)").arg(supportedImageFormatsString));
+}
+
+QString YACReader::imagePathFromMimeData(const QMimeData *mimeData)
+{
+    QString filePath;
+
+    if (mimeData->hasUrls()) {
+        QList<QUrl> urlList = mimeData->urls();
+
+        if (!urlList.isEmpty()) {
+            QUrl url = urlList.first();
+            if (url.isLocalFile()) {
+                filePath = url.toLocalFile();
+
+                QFileInfo fileInfo(filePath);
+                QString extension = fileInfo.suffix().toLower();
+                QList<QByteArray> supportedFormats = QImageReader::supportedImageFormats();
+                bool isSupported = false;
+
+                for (const QByteArray &format : supportedFormats) {
+                    if (extension == QString(format).toLower()) {
+                        isSupported = true;
+                        break;
+                    }
+                }
+
+                if (!isSupported) {
+                    filePath.clear();
+                }
+            }
+        }
+    }
+
+    return filePath;
 }