+++ common/yacreader_global_gui.h
@@ -112,6 +112,7 @@ QString addExtensionToIconPath(const QString &path);
 QString addExtensionToIconPathInToolbar(const QString &path);
 QAction *actionWithCustomIcon(const QIcon &icon, QAction *action);
 QPixmap hdpiPixmap(const QString &file, QSize size);
-
+QString imageFileLoader(QWidget *parent);
+QString imagePathFromMimeData(const QMimeData *mimeData);
 }
 #endif